tss: add tpm2_PolicySecret
Add a cut down version of PolicySecret (with no nonceTPM, cpHashA or
expiry and thus no returned timeout or ticket) for use with a
PolicySecret statement in the encoded policy statements.

For simplicity the authorization string is passed in and the internal
routine constructs the necessary HMAC session for encoding it in the
command (this saves the calling code from having to worry about
multiple sessions).

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>