doc: clarify use of the secret parameter
In the original OID allocation for TPM keys, there is a separate OID
for a loadable and an importable key.  However, as the standard
evolved it was realised that simply the presence or absence of the
secret optional parameter is sufficient to distinguish the key types,
so for sealed keys.  Update the secret section to make this distinction.

Reviewed-by: Morten Linderud <morten@linderud.pw>
Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>