Fix bug in handling wrapping of the secret component
The spec says we should insert the entire marshalled
TPM2B_ENCRYPTED_SECRET as the ASN.1 secret component, however we are
marshalling only the secret part not the initial length.  However, we
are unmarshalling a TPM2B_ENCRYPTED_SECRET correctly.  This works
because there's a bug in the way we marshal the elliptic curve public
point into the secret->secret.  We're actually marshalling it as
TPM2B_ECC_POINT instead of what it should be, a TPMS_ECC_POINT.  This
accidentally makes secret->secret a fully marshalled
TPM2B_ENCRYPTED_SECRET and so we obey the actual letter of the
standard (albeit accidentally).

Fix this by correctly marshalling the TPM2B_ENCRYPTED_SECRET and
putting the correct TPMS_ECC_POINT into the secret->secret area.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>