libcommon: SECURITY: always salt sessions
THe problem with unsalted sessions is that the session key is derived
purely from two observable nonces and the authority.  This means that
in the case of no authority (policy with no password) the encryption
key is trivially derivable.  However even with a password, the session
key is fast to compute meaning an observer who captures the nonces,
the command and its hmac can run a very fast offline brute force
attempt to crack the password.  Salting the session defeats this
because the unencrypted salt value is also used to derive the session
key.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>